+++ src/source/tuner/TunerModel.java
@@ -220,7 +220,7 @@ public class TunerModel extends AbstractTableModel implements Listener<TunerEven
                 case TUNER_ID:
                     return tuner.getUniqueID();
                 case SAMPLE_RATE:
-                    int sampleRate = tuner.getTunerController().getSampleRate();
+                    double sampleRate = tuner.getTunerController().getSampleRate();
 
                     return mSampleRateFormat.format(sampleRate / 1E6D) + MHZ;
                 case FREQUENCY:
